# Backend/api_documentation.yaml
openapi: 3.1.0
info:
  title: CivicResolve AI
  version: 0.1.0
paths:
...
  /api/citizen/facility/{facility_id}:
    get:
      tags:
      - citizen
      summary: Citizen Facility Detail
      operationId: citizen_facility_detail_api_citizen_facility__facility_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: facility_id
        in: path
        required: true
        schema:
          type: integer
          title: Facility Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitizenFacilityDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/commissioner/dash:
    get:
      tags:
      - commissioner
      summary: Commissioner Dashboard
      operationId: commissioner_dashboard_api_commissioner_dash_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissionerDashboardResponse'
      security:
      - HTTPBearer: []
...
  /api/commissioner/officers:
    get:
      tags:
      - commissioner
      summary: Commissioner Officers List
      operationId: commissioner_officers_list_api_commissioner_officers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/UserSchema'
                  type: array
                type: object
                title: Response Commissioner Officers List Api Commissioner Officers
                  Get
      security:
      - HTTPBearer: []
...
  /api/commissioner/facility:
    post:
      tags:
      - commissioner
      summary: Commissioner Add Facility
      operationId: commissioner_add_facility_api_commissioner_facility_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionerFacilityRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Commissioner Add Facility Api Commissioner Facility
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/commissioner/facility/{facility_id}:
    put:
      tags:
      - commissioner
      summary: Commissioner Update Facility
      operationId: commissioner_update_facility_api_commissioner_facility__facility_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: facility_id
        in: path
        required: true
        schema:
          type: integer
          title: Facility Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionerFacilityRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Commissioner Update Facility Api Commissioner Facility  Facility
                  Id  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - commissioner
      summary: Commissioner Delete Facility
      operationId: commissioner_delete_facility_api_commissioner_facility__facility_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: facility_id
        in: path
        required: true
        schema:
          type: integer
          title: Facility Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Commissioner Delete Facility Api Commissioner Facility  Facility
                  Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/commissioner/facility-type/{type_id}:
    put:
      tags:
      - commissioner
      summary: Update Commissioner Facility Type
      operationId: update_commissioner_facility_type_api_commissioner_facility_type__type_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: type_id
        in: path
        required: true
        schema:
          type: integer
          title: Type Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionerFacilityTypeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissionerFacilityTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - commissioner
      summary: Delete Commissioner Facility Type
      operationId: delete_commissioner_facility_type_api_commissioner_facility_type__type_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: type_id
        in: path
        required: true
        schema:
          type: integer
          title: Type Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/commissioner/citizens:
    get:
      tags:
      - commissioner
      summary: Commissioner Citizens List
      operationId: commissioner_citizens_list_api_commissioner_citizens_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/UserSchema'
                  type: array
                type: object
                title: Response Commissioner Citizens List Api Commissioner Citizens
                  Get
      security:
      - HTTPBearer: []
...
  /api/commissioner/profile:
    get:
      tags:
      - commissioner
      summary: Commissioner Profile Fetch
      operationId: commissioner_profile_fetch_api_commissioner_profile_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    CitizenFacilityDetailResponse:
      properties:
        facility:
          $ref: '#/components/schemas/FacilitySchema'
        bookedDates:
          items:
            type: string
          type: array
          title: Bookeddates
        myBookedDates:
          items:
            type: string
          type: array
          title: Mybookeddates
      type: object
      required:
      - facility
      - bookedDates
      - myBookedDates
      title: CitizenFacilityDetailResponse
...
    CommissionerDashboardResponse:
      properties:
        totalComplaints:
          type: integer
          title: Totalcomplaints
        pendingComplaints:
          type: integer
          title: Pendingcomplaints
        resolvedComplaints:
          type: integer
          title: Resolvedcomplaints
        closedComplaints:
          type: integer
          title: Closedcomplaints
        criticalComplaints:
          type: integer
          title: Criticalcomplaints
        totalOfficers:
          type: integer
          title: Totalofficers
        totalCitizens:
          type: integer
          title: Totalcitizens
        totalRevenue:
          type: number
          title: Totalrevenue
        billRevenue:
          type: number
          title: Billrevenue
        bookingRevenue:
          type: number
          title: Bookingrevenue
        complaintsByCategory:
          items:
            $ref: '#/components/schemas/NameCount'
          type: array
          title: Complaintsbycategory
        complaintsByStatus:
          items:
            $ref: '#/components/schemas/StatusCount'
          type: array
          title: Complaintsbystatus
        trend:
          items:
            $ref: '#/components/schemas/TrendData'
          type: array
          title: Trend
      type: object
      required:
      - totalComplaints
      - pendingComplaints
      - resolvedComplaints
      - closedComplaints
      - criticalComplaints
      - totalOfficers
      - totalCitizens
      - totalRevenue
      - billRevenue
      - bookingRevenue
      - complaintsByCategory
      - complaintsByStatus
      - trend
      title: CommissionerDashboardResponse
    CommissionerFacilityRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        facilityType:
          anyOf:
          - type: string
          - type: 'null'
          title: Facilitytype
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        pincode:
          anyOf:
          - type: string
          - type: 'null'
          title: Pincode
        pricePerDay:
          anyOf:
          - type: string
          - type: number
          - type: 'null'
          title: Priceperday
        capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Capacity
        amenities:
          anyOf:
          - {}
          - type: 'null'
          title: Amenities
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        isActive:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isactive
      type: object
      title: CommissionerFacilityRequest
    CommissionerFacilityTypeRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      title: CommissionerFacilityTypeRequest
    CommissionerFacilityTypeResponse:
      properties:
        message:
          type: string
          title: Message
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - message
      - id
      - name
      title: CommissionerFacilityTypeResponse
...
    FacilitySchema:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        facilityType:
          type: string
          title: Facilitytype
        address:
          type: string
          title: Address
        pincode:
          type: string
          title: Pincode
        pricePerDay:
          type: number
          title: Priceperday
        capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Capacity
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        amenities:
          anyOf:
          - {}
          - type: 'null'
          title: Amenities
        isActive:
          type: boolean
          title: Isactive
      type: object
      required:
      - id
      - name
      - facilityType
      - address
      - pincode
      - pricePerDay
      - isActive
      title: FacilitySchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NameCount:
      properties:
        name:
          type: string
          title: Name
        count:
          type: integer
          title: Count
      type: object
      required:
      - name
      - count
      title: NameCount
...
    StatusCount:
      properties:
        status:
          type: string
          title: Status
        count:
          type: integer
          title: Count
      type: object
      required:
      - status
      - count
      title: StatusCount
...
    TrendData:
      properties:
        day:
          type: string
          title: Day
        filed:
          type: integer
          title: Filed
        resolved:
          type: integer
          title: Resolved
      type: object
      required:
      - day
      - filed
      - resolved
      title: TrendData
    UserSchema:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          title: Email
        name:
          type: string
          title: Name
        role:
          type: string
          title: Role
        badgeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Badgeid
        departmentId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Departmentid
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        pincode:
          anyOf:
          - type: string
          - type: 'null'
          title: Pincode
        isActive:
          type: boolean
          title: Isactive
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      type: object
      required:
      - id
      - email
      - name
      - role
      - isActive
      title: UserSchema
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer